Move classes in test_tools to conftest
