+++ tests/processes/wps_test_process.py
@@ -16,10 +16,7 @@ class TestProcess(Process):
         }
         inputs = [
             LiteralInput(
-                "string",
-                "String",
-                abstract="Enter string.",
-                data_type="string",
+                "string", "String", abstract="Enter string.", data_type="string",
             )
         ]
         outputs = [
